data processing for shakespeare sonnets done
