+++ ShakespeareDeRedaction/sonnets_dataset/dataset_dict.json
@@ -1,0 +1,1 @@
+{"splits": ["train", "test", "validation"]}
