+++ ShakespeareDeRedaction/sonnets_dataset/validation/dataset_info.json
@@ -1,0 +1,12 @@
+{
+  "citation": "",
+  "description": "",
+  "features": {
+    "text": {
+      "dtype": "string",
+      "_type": "Value"
+    }
+  },
+  "homepage": "",
+  "license": ""
+}
